Refactored post processing into separate components
